Refactor ResumenReservacion
