+++ DREAM Lab Frontend/src/pages/ResumenReservacion/assets/images/greenTickIcon.png
